Move the reexec parts of lib/srv to session/reexec
