+++ lib/srv/exec.go
@@ -19,8 +19,8 @@
 package srv
 
 import (
-	"bufio"
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -43,16 +43,8 @@ import (
 	apievents "github.com/gravitational/teleport/api/types/events"
 	"github.com/gravitational/teleport/lib/events"
 	"github.com/gravitational/teleport/lib/sshutils/reexec"
-	"github.com/gravitational/teleport/lib/utils"
-)
-
-const (
-	defaultPath          = "/bin:/usr/bin:/usr/local/bin:/sbin"
-	defaultEnvPath       = "PATH=" + defaultPath
-	defaultRootPath      = "/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
-	defaultEnvRootPath   = "PATH=" + defaultRootPath
-	defaultTerm          = "xterm"
-	defaultLoginDefsPath = "/etc/login.defs"
+	"github.com/gravitational/teleport/session/envutils"
+	sessionreexec "github.com/gravitational/teleport/session/reexec"
 )
 
 // ExecResult is used internally to send the result of a command execution from
@@ -591,66 +583,6 @@ func emitExecAuditEvent(ctx *ServerContext, cmd string, execErr error) {
 	}
 }
 
-// getDefaultEnvPath returns the default value of PATH environment variable for
-// new logins (prior to shell) based on login.defs. Returns a string which
-// looks like "PATH=/usr/bin:/bin"
-func getDefaultEnvPath(uid string, loginDefsPath string) string {
-	envPath := defaultEnvPath
-	envRootPath := defaultEnvRootPath
-
-	// open file, if it doesn't exist return a default path and move on
-	f, err := utils.OpenFileAllowingUnsafeLinks(loginDefsPath)
-	if err != nil {
-		if uid == "0" {
-			slog.DebugContext(context.Background(), "Unable to open login.defs, returning default su path", "login_defs_path", loginDefsPath, "error", err, "default_path", defaultEnvRootPath)
-			return defaultEnvRootPath
-		}
-		slog.DebugContext(context.Background(), "Unable to open login.defs, returning default path", "login_defs_path", loginDefsPath, "error", err, "default_path", defaultEnvPath)
-		return defaultEnvPath
-	}
-	defer f.Close()
-
-	// read path from login.defs file (/etc/login.defs) line by line:
-	scanner := bufio.NewScanner(f)
-	for scanner.Scan() {
-		line := strings.TrimSpace(scanner.Text())
-
-		// skip comments and empty lines:
-		if line == "" || line[0] == '#' {
-			continue
-		}
-
-		// look for a line that starts with ENV_PATH or ENV_SUPATH
-		fields := strings.Fields(line)
-		if len(fields) > 1 {
-			if fields[0] == "ENV_PATH" {
-				envPath = fields[1]
-			}
-			if fields[0] == "ENV_SUPATH" {
-				envRootPath = fields[1]
-			}
-		}
-	}
-
-	// if any error occurs while reading the file, return the default value
-	err = scanner.Err()
-	if err != nil {
-		if uid == "0" {
-			slog.WarnContext(context.Background(), "Unable to read login.defs, returning default su path", "login_defs_path", loginDefsPath, "error", err, "default_path", defaultEnvRootPath)
-			return defaultEnvRootPath
-		}
-		slog.WarnContext(context.Background(), "Unable to read login.defs, returning default path", "login_defs_path", loginDefsPath, "error", err, "default_path", defaultEnvPath)
-		return defaultEnvPath
-	}
-
-	// if requesting path for uid 0 and no ENV_SUPATH is given, fallback to
-	// ENV_PATH first, then the default path.
-	if uid == "0" {
-		return envRootPath
-	}
-	return envPath
-}
-
 // parseSecureCopy will parse a command and return if it's secure copy or not.
 func parseSecureCopy(path string) (string, string, bool, error) {
 	parts := strings.Fields(path)
@@ -707,5 +639,97 @@ func exitCode(err error) int {
 	default:
 		slog.DebugContext(context.Background(), "Unknown error returned when executing command", "error", err)
 		return teleport.RemoteCommandFailure
+	}
+}
+
+// ConfigureCommand creates a command fully configured to execute. This
+// function is used by Teleport to re-execute itself and pass whatever data
+// is need to the child to actually execute the shell.
+func ConfigureCommand(ctx *ServerContext, extraFiles ...*os.File) (*exec.Cmd, error) {
+	// Create a os.Pipe and start copying over the payload to execute. While the
+	// pipe buffer is quite large (64k) some users have run into the pipe
+	// blocking writes on much smaller buffers (7k) leading to Teleport being
+	// unable to run some exec commands.
+	//
+	// To not depend on the OS implementation of a pipe, instead the copy should
+	// be non-blocking. The io.Copy will be closed when either when the child
+	// process has fully read in the payload or the process exits with an error
+	// (and closes all child file descriptors).
+	//
+	// See the below for details.
+	//
+	//   https://man7.org/linux/man-pages/man7/pipe.7.html
+	cmdmsg, err := ctx.ExecCommand()
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	go copyCommand(ctx.CancelContext(), ctx.cmdw, cmdmsg)
+
+	// Find the Teleport executable and its directory on disk.
+	executable, err := os.Executable()
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	// The channel/request type determines the subcommand to execute.
+	var subCommand string
+	switch ctx.ExecType {
+	case teleport.NetworkingSubCommand:
+		subCommand = teleport.NetworkingSubCommand
+	default:
+		subCommand = teleport.ExecSubCommand
+	}
+
+	// Build the list of arguments to have Teleport re-exec itself. The "-d" flag
+	// is appended if Teleport is running in debug mode.
+	args := []string{executable, subCommand}
+
+	// build env for `teleport exec`
+	env := &envutils.SafeEnv{}
+	env.AddExecEnvironment()
+
+	// Build the "teleport exec" command.
+	cmd := &exec.Cmd{
+		Path: executable,
+		Args: args,
+		Env:  *env,
+		ExtraFiles: []*os.File{
+			ctx.cmdr,
+			ctx.logw,
+			ctx.contr,
+			ctx.readyw,
+			ctx.killShellr,
+		},
+	}
+	// Add extra files if applicable.
+	if len(extraFiles) > 0 {
+		cmd.ExtraFiles = append(cmd.ExtraFiles, extraFiles...)
+	}
+
+	// Perform OS-specific tweaks to the command.
+	sessionreexec.CommandOSTweaks(cmd)
+
+	return cmd, nil
+}
+
+// copyCommand will copy the provided command to the child process over the
+// pipe attached to the context.
+func copyCommand(ctx context.Context, cmdw *os.File, cmdmsg *sessionreexec.ExecCommand) {
+	defer func() {
+		err := cmdw.Close()
+		if err != nil {
+			slog.ErrorContext(ctx, "Failed to close command pipe", "error", err)
+		}
+
+		// Set to nil so the close in the context doesn't attempt to re-close.
+		cmdw = nil
+	}()
+
+	// Write command bytes to pipe. The child process will read the command
+	// to execute from this pipe.
+	if err := json.NewEncoder(cmdw).Encode(cmdmsg); err != nil {
+		slog.ErrorContext(ctx, "Failed to copy command over pipe", "error", err)
+		return
 	}
 }
