+++ lib/srv/reexec_inline.go
@@ -1,0 +1,41 @@
+// Teleport
+// Copyright (C) 2026 Gravitational, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+package srv
+
+import (
+	"os/user"
+
+	"github.com/gravitational/teleport/session/reexec"
+)
+
+//go:fix inline
+func IsReexec() bool {
+	return reexec.IsReexec()
+}
+
+//go:fix inline
+func RunAndExit(commandType string) {
+	reexec.RunAndExit(commandType)
+}
+
+//go:fix inline
+type ExecLogConfig = reexec.ExecLogConfig
+
+//go:fix inline
+func CheckHomeDir(localUser *user.User) (bool, error) {
+	return reexec.CheckHomeDir(localUser)
+}
